+++ includes/connectionManagement.php
@@ -92,7 +92,7 @@
     public static function loadConnectionFile($file, $connect = false) {
       // Parse the files using an ini parser.
       $connection = parse_ini_file($file, true);
-      Logger::debug(var_export($connection, true));
+      Logger::devel(var_export($connection, true));
 
       // Require these items to be defined.
       if (isset($connection['address']) && isset($connection['port'])) {
@@ -120,7 +120,7 @@
       }
       else {
         // Uh-oh!
-        Logger::debug("Connection in file \"".$file."\" failed to parse.");
+        Logger::info("Connection in file \"".$file."\" failed to parse.");
       }
       return false;
     }