+++ main.php
@@ -1,44 +1,80 @@
 <?php
   class Main {
     public function __construct($argv) {
-      // Verify that the bot can run in the provided environment.
+      // Verify that the bot can run in the provided environment
       $this->verifyEnvironment();
 
       if (isset($argv[1])) {
-        // Ensure that any non-default user input is converted to a boolean.
-        $debug = (bool)$argv[1];
+        // Ensure that any non-default user input is converted to an integer
+        $loglevel = (int)$argv[1];
+      }
+      elseif (is_readable(__PROJECTROOT__."/conf/loglevel.conf")) {
+        // Read the log level from conf/loglevel.conf
+        $loglevel = (int)file_get_contents(__PROJECTROOT__.
+          "/conf/loglevel.conf");
       }
       else {
-        $debug = false;
+        $loglevel = 0;
       }
 
-      // Activate full error reporting.
+      // Activate full error reporting
       $this->setErrorReporting();
 
-      // Setup required constants for operation and load required classes.
-      $this->prepareEnvironment($debug);
+      // Setup required constants for operation and load required classes
+      $this->prepareEnvironment($loglevel);
 
-      // Brag about versions and junk.
+      // Brag about versions and junk
       $this->brag();
 
-      // Load requested modules.
+      // Load requested modules
       $this->loadModules();
 
-      // Discover sockets located in conf/listen.conf.
+      // Discover sockets located in conf/listen.conf
       $this->discoverSockets();
 
-      // Discover connections located in conf/connections/.
+      // Background the process (if we can/should)
+      $this->background();
+
+      // Discover connections located in conf/connections/
       $this->discoverConnections();
 
-      // Initiate all loaded connections.
+      // Initiate all loaded connections
       $this->activateConnections();
     }
 
     private function activateConnections() {
-      // Iterate through the list of defined connections.
+      // Iterate through the list of defined connections
       foreach (ConnectionManagement::getConnections() as $connection) {
-        // Connect.
+        // Connect
         $connection->connect();
+      }
+    }
+
+    private function background() {
+      // Only background if we're in silent mode and have PCNTL
+      if (__LOGLEVEL__ == 0 && function_exists("pcntl_fork")
+          && function_exists("pcntl_fork")) {
+        if ($pid = pcntl_fork()) {
+          die();
+        }
+
+        // Discard the output buffer and close
+        ob_end_clean();
+
+        // Close all of the standard file descriptors
+        fclose(STDIN);
+        fclose(STDOUT);
+        fclose(STDERR);
+
+        register_shutdown_function(
+          function() {
+            posix_kill(posix_getpid(), SIGHUP);
+          }
+        );
+
+        if (posix_setsid() < 0 || $pid = pcntl_fork()) {
+          die();
+        }
       }
     }
 
@@ -46,7 +82,7 @@
       Logger::info("Welcome to Modfwango!");
       Logger::info("You're running Modfwango v".
         __MODFWANGOVERSION__.".");
-      // Check for updates to Modfwango.
+      // Check for updates to Modfwango
       $version = "1.00";
       $contents = @explode("\n", @file_get_contents("https://raw.githubusercon".
         "tent.com/Modfwango/Modfwango/master/docs/CHANGELOG.md", 0,
@@ -66,33 +102,33 @@
     }
 
     private function discoverConnections() {
-      // Get a list of connection configurations.
+      // Get a list of connection configurations
       $connections = glob(__PROJECTROOT__."/conf/connections/*.conf");
 
-      // Iterate through the list and load each item individually.
+      // Iterate through the list and load each item individually
       foreach ($connections as $file) {
         ConnectionManagement::loadConnectionFile($file);
       }
     }
 
     private function discoverSockets() {
-      // Load the listen configuration.
+      // Load the listen configuration
       $listen = trim(file_get_contents(__PROJECTROOT__."/conf/listen.conf"));
       $listen = explode("\n", $listen);
-      // Iterate through each line.
+      // Iterate through each line
       foreach ($listen as $sock) {
-        // Make sure there are no stray line endings.
+        // Make sure there are no stray line endings
         $sock = trim($sock);
-        // Make sure the line has a non-null value and a comma.
+        // Make sure the line has a non-null value and a comma
         if (strlen($sock) > 0 && strstr($sock, ",")) {
-          // Separate bind address from bind port.
+          // Separate bind address from bind port
           $sock = explode(",", $sock);
-          // Make sure we have the correct amount of parameters.
+          // Make sure we have the correct amount of parameters
           if (count($sock) == 2) {
-            // Attempt to create a socket.
+            // Attempt to create a socket
             $sock = new Socket(trim($sock[0]), trim($sock[1]));
             if ($sock != false) {
-              // Add it to the socket management class.
+              // Add it to the socket management class
               SocketManagement::newSocket($sock);
             }
             else {
@@ -103,20 +139,22 @@
         }
       }
 
-      // Attempt to create the inter-process communication socket.
-      $sock = new Socket("127.0.0.1", "0", true);
-      if ($sock != false) {
-        // Add it to the socket management class.
-        SocketManagement::newSocket($sock);
-      }
-      else {
-        // Couldn't bind!
-        Logger::debug("Could not bind to address.");
+      if (function_exists("pcntl_fork")) {
+        // Attempt to create the inter-process communication socket
+        $sock = new Socket("127.0.0.1", "0", true);
+        if ($sock != false) {
+          // Add it to the socket management class
+          SocketManagement::newSocket($sock);
+        }
+        else {
+          // Couldn't bind!
+          Logger::debug("Could not bind to address.");
+        }
       }
     }
 
     private function loadModules() {
-      // Load modules in requested order in modfwango conf/modules.conf.
+      // Load modules in requested order in modfwango conf/modules.conf
       foreach (explode("\n",
           trim(file_get_contents(__MODFWANGOROOT__."/conf/modules.conf")))
           as $module) {
@@ -130,7 +168,7 @@
         }
       }
 
-      // Load modules in requested order in project conf/modules.conf.
+      // Load modules in requested order in project conf/modules.conf
       foreach (explode("\n",
           trim(file_get_contents(__PROJECTROOT__."/conf/modules.conf")))
           as $module) {
@@ -144,7 +182,7 @@
         }
       }
 
-      // Make sure all requested modules were loaded.
+      // Make sure all requested modules were loaded
       foreach ($modules as $module) {
         if (!ModuleManagement::isLoaded(basename($module))) {
           Logger::info("Module \"".$module."\" failed to load.");
@@ -154,67 +192,69 @@
     }
 
     public function loop() {
-      // Infinitely loop.
+      // Infinitely loop
       while (true) {
-        // Iterate through each socket.
+        // Iterate through each socket
         foreach (SocketManagement::getSockets() as $socket) {
-          // Attempt to accept new connections.
+          // Attempt to accept new connections
           $socket->accept();
         }
 
-        // Prune dead connections.
+        // Prune dead connections
         ConnectionManagement::pruneConnections();
 
-        // Iterate through each connection.
+        // Iterate through each connection
         foreach (ConnectionManagement::getConnections() as $connection) {
-          // Fetch any received data.
+          // Fetch any received data
           $data = trim($connection->getData());
           if ($data != false) {
             if (stristr($data, "\n")) {
               foreach (explode("\n", $data) as $line) {
-                if ($connection->getIPC() == true) {
-                  // Pass the connection and associated data to the IPC handler.
+                if (function_exists("pcntl_fork")
+                    && $connection->getIPC() == true) {
+                  // Pass the connection and associated data to the IPC handler
                   IPCHandling::receiveData($connection, trim($line));
                 }
                 else {
                   // Pass the connection and associated data to the event
-                  // handler.
+                  // handler
                   EventHandling::receiveData($connection, trim($line));
                 }
               }
             }
             else {
-              if ($connection->getIPC() == true) {
-                // Pass the connection and associated data to the IPC handler.
+              if (function_exists("pcntl_fork")
+                  && $connection->getIPC() == true) {
+                // Pass the connection and associated data to the IPC handler
                 IPCHandling::receiveData($connection, trim($data));
               }
               else {
-                // Pass the connection and associated data to the event handler.
+                // Pass the connection and associated data to the event handler
                 EventHandling::receiveData($connection, trim($data));
               }
             }
           }
         }
 
-        // Get the connectionLoopEndEvent event.
+        // Get the connectionLoopEndEvent event
         $event = EventHandling::getEventByName("connectionLoopEndEvent");
         if ($event != false) {
           foreach ($event[2] as $id => $registration) {
             // Trigger the connectionLoopEndEvent event for each registered
-            // module.
+            // module
             EventHandling::triggerEvent("connectionLoopEndEvent", $id);
           }
         }
-        // Sleep for a small amount of time to prevent high CPU usage.
+        // Sleep for a small amount of time to prevent high CPU usage
         usleep(__DELAY__);
       }
     }
 
-    private function prepareEnvironment($debug) {
-      // Define the root of the Modfwango library folder.
+    private function prepareEnvironment($loglevel) {
+      // Define the root of the Modfwango library folder
       define("__MODFWANGOROOT__", dirname(__FILE__));
 
-      // Locate the latest version in docs/CHANGELOG.md.
+      // Locate the latest version in docs/CHANGELOG.md
       $version = "1.00";
       if (file_exists(__MODFWANGOROOT__."/docs/CHANGELOG.md")) {
         $contents = explode("\n", file_get_contents(__MODFWANGOROOT__.
@@ -228,43 +268,46 @@
         }
       }
 
-      // Define the current version of Modfwango.
+      // Define the current version of Modfwango
       define("__MODFWANGOVERSION__", $version);
 
-      // Change current working directory to project root.
+      // Change current working directory to project root
       chdir(__PROJECTROOT__);
 
-      // Set the default timezone.
+      // Set the default timezone
       if (defined("__TIMEZONE__")) {
         date_default_timezone_set(__TIMEZONE__);
       }
 
-      // Define start timestamp.
+      // Define start timestamp
       define("__STARTTIME__", time());
 
-      // Define the debug constant to allow the logger determine the correct
-      // output type.
-      define("__DEBUG__", $debug);
-
-      // Define the time to sleep at the end of every infinite loop.
+      // Define the time to sleep at the end of every infinite loop
       define("__DELAY__", 10000);
 
-      // Load the connection related classes.
+      // Define the debug constant to allow the logger determine the correct
+      // output type
+      define("__LOGLEVEL__", $loglevel);
+
+      // Load the connection related classes
       require_once(__MODFWANGOROOT__."/includes/connection.php");
       require_once(__MODFWANGOROOT__."/includes/connectionManagement.php");
       require_once(__MODFWANGOROOT__."/includes/socket.php");
       require_once(__MODFWANGOROOT__."/includes/socketManagement.php");
 
-      // Load the inter-process communication handler.
-      require_once(__MODFWANGOROOT__."/includes/IPCHandling.php");
+      if (function_exists("pcntl_fork")) {
+        Logger::debug("PCNTL support isn't available.");
+        // Load the inter-process communication handler
+        require_once(__MODFWANGOROOT__."/includes/IPCHandling.php");
+      }
 
-      // Load the event handler.
+      // Load the event handler
       require_once(__MODFWANGOROOT__."/includes/eventHandling.php");
 
-      // Load the logger.
+      // Load the logger
       require_once(__MODFWANGOROOT__."/includes/logger.php");
 
-      // Make sure the launcher is up-to-date.
+      // Make sure the launcher is up-to-date
       if (!file_exists(__PROJECTROOT__."/main.php")
           || hash("md5", file_get_contents(__MODFWANGOROOT__."/launcher.php"))
           != hash("md5", file_get_contents(__PROJECTROOT__."/main.php"))) {
@@ -273,19 +316,19 @@
         Logger::info("The launcher has been updated.");
       }
 
-      // Load the module management class.
+      // Load the module management class
       require_once(__MODFWANGOROOT__."/includes/moduleManagement.php");
 
-      // Load the storage handling class.
+      // Load the storage handling class
       require_once(__MODFWANGOROOT__."/includes/storageHandling.php");
 
-      // Register signal handlers.
+      // Register signal handlers
       declare(ticks = 1);
       if (function_exists("pcntl_signal")) {
         pcntl_signal(SIGINT,
           function() {
             echo "\r";
-            Logger::info("Caught SIGINT; Begin shutdown procedure...");
+            Logger::info("Begin shutdown procedure...");
             foreach (ConnectionManagement::getConnections() as $c) {
               $c->disconnect();
             }
@@ -294,12 +337,6 @@
             }
             Logger::info("Shutting down...");
             die();
-          }
-        );
-        pcntl_signal(SIGTSTP,
-          function() {
-            echo "\r";
-            Logger::memoryUsage();
           }
         );
       }
@@ -311,26 +348,26 @@
     }
 
     private function verifyEnvironment() {
-      // Verify that the current directory structure is named safely.
+      // Verify that the current directory structure is named safely
       if (!preg_match("/^[a-zA-Z0-9\\/.\\-]+$/", dirname(__FILE__))) {
         die("The full path to this file must match this regular expression:\n^".
           "[a-zA-Z0-9\\/.\\-]+$\n");
       }
 
-      // Verify that the launcher script has setup the project root constant.
+      // Verify that the launcher script has setup the project root constant
       if (!defined("__PROJECTROOT__")) {
         die("__PROJECTROOT__ hasn't been defined by a launcher script.\n");
       }
     }
   }
 
-  // Instantiate the bot to get things moving.
+  // Instantiate the bot to get things moving
   $main = new Main($argv);
 
-  // Start the main loop.
+  // Start the main loop
   $main->loop();
 
-  // Allow things to easily get the main class.
+  // Allow things to easily get the main class
   function getMain() {
     return $GLOBALS['main'];
   }