+++ modules/libraries/Timer.php
@@ -22,7 +22,9 @@
             if (is_object($mod)) {
               // Make sure both classnames match.
               if (get_class($mod) == get_class($class)) {
+                Logger::stack("Entering module: ".$class->name."::".$callback);
                 $class->$callback($timer["params"]);
+                Logger::stack("Left module: ".$class->name."::".$callback);
               }
               else {
                 // Attempted to use an old module.
@@ -34,7 +36,9 @@
           }
           else {
             if (is_object($class)) {
+              Logger::stack("Entering module: ".$class->name."::".$callback);
               $class->$callback($timer["params"]);
+              Logger::stack("Left module: ".$class->name."::".$callback);
             }
           }